04-multimodal/requirements.txt
aiogram
openai
python-dotenv
pydantic
ollama
aiohttp


